Adding RAdding LiquidPropertyRegistrar instead of tag
