+++ Lombiq.Tests.UI.AppExtensions/Lombiq.Tests.UI.AppExtensions.csproj
@@ -41,8 +41,4 @@
     <ProjectReference Include="..\..\..\src\Libraries\Lombiq.HelpfulLibraries\Lombiq.HelpfulLibraries.OrchardCore\Lombiq.HelpfulLibraries.OrchardCore.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="ParserTags\" />
-  </ItemGroup>
-
 </Project>
