--- Lombiq.Tests.UI.AppExtensions/ParserTags/IsUITestingLiquidParserTag.cs
@@ -1,30 +1,0 @@
-using Fluid;
-using Fluid.Ast;
-using Lombiq.HelpfulLibraries.OrchardCore.Liquid;
-using Microsoft.Extensions.Configuration;
-using System.Collections.Generic;
-using System.IO;
-using System.Text.Encodings.Web;
-using System.Threading.Tasks;
-
-namespace Lombiq.Tests.UI.AppExtensions.ParserTags;
-
-public class IsUITestingLiquidParserTag : ILiquidParserTag
-{
-    private readonly IConfiguration _configuration;
-
-    public IsUITestingLiquidParserTag(IConfiguration configuration) =>
-        _configuration = configuration;
-
-    public async ValueTask<Completion> WriteToAsync(
-        IReadOnlyList<FilterArgument> argumentsList,
-        TextWriter writer,
-        TextEncoder encoder,
-        TemplateContext context)
-    {
-        var isUiTesting = _configuration.IsUITesting();
-
-        await writer.WriteAsync(isUiTesting ? "true" : "false");
-        return Completion.Normal;
-    }
-}
